Add userID feature to categories

+++ lib/app/features/home/drink_page/repository/drink_documents_repository.dart
@@ -1,9 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:moja_lodowka/app/features/home/drink_page/model/drink_document_model.dart';
 
 class DrinkDocumentsRepository {
   Stream<List<DrinkDocumentModel>> getDocumentsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('napoje')
         .orderBy('title')
         .snapshots()
@@ -18,13 +25,25 @@ class DrinkDocumentsRepository {
   }
 
   Future<void> add(String title, DateTime expDate) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('napoje')
         .add({'title': title, 'expdate': expDate});
   }
 
   Future<void> delete({required String document}) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('napoje')
         .doc(document)
         .delete();
