+++ lib/app/features/home/fridge_page/repository/fridge_documents_repository.dart
@@ -1,9 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:moja_lodowka/app/features/home/fridge_page/model/fridge_document_model.dart';
 
 class FridgeDocumentsRepository {
   Stream<List<FridgeDocumentModel>> getDocumentsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('lodowka')
         .orderBy('title')
         .snapshots()
@@ -19,13 +26,25 @@ class FridgeDocumentsRepository {
   }
 
   Future<void> add(String title, DateTime expDate) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('lodowka')
         .add({'title': title, 'expdate': expDate});
   }
 
   Future<void> delete({required String document}) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Użytkownik niezalogowany');
+    }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('lodowka')
         .doc(document)
         .delete();
